Make coordinates precision configurable.
The PolylineEncoder class used to use a fixed decimal places precision
(5 digits) and this conformed with Google's utility which rounded
all values up to the fifth digit. This change makes the number of
digits configurable. The PolylineEncoder is now a generic class whose
template parameter defines the number of decimal places it should use
(defaulted to five).